Added authorization and authentication.

+++ src/main/java/com/hagz_hotels/hotels_booking/Controllers/Admin/DeleteRoomService.java
@@ -1,7 +1,10 @@
 package com.hagz_hotels.hotels_booking.Controllers.Admin;
 
 
+import com.hagz_hotels.hotels_booking.Controllers.Auth;
 import com.hagz_hotels.hotels_booking.Model.DAO.RoomDAO;
+import com.hagz_hotels.hotels_booking.Model.Entities.User;
+import com.hagz_hotels.hotels_booking.Util.JsonResponse;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -15,7 +18,7 @@ import java.io.IOException;
 public class DeleteRoomService extends HttpServlet {
 
     RoomDAO roomDAO;
-
+    User.Type authType = User.Type.ADMIN;
     @Override
     public void init() throws ServletException {
         roomDAO = new RoomDAO();
@@ -23,8 +26,16 @@ public class DeleteRoomService extends HttpServlet {
 
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        // TODO: Authentication and return error
+        response.setContentType("application/json");
+        if (!Auth.authenticateJson(request, response, authType))
+            return;
+
         Integer roomId = Integer.valueOf(request.getParameter("roomId"));
+
+        if (!Auth.authorizeJsonRoom(request, response, roomId))
+            return;
+        JsonResponse jsonResponse = new JsonResponse();
         roomDAO.delete(roomId);
+        response.getWriter().println(jsonResponse);
     }
 }
